Include the Still Capture StreamRole
I've only set the still image capture to happen once at the start and I
still need to hook it up to the shutter button, but in theory, I can
take still images whilst using the viewfinder?  I should also consider
turning off the display when the still is being captured - to avoid
light leaking in.

Oh and I'd like to have a "red" display mode and allow for setting the
exposure time, auto-focus on/off, and ISO setting. (The lens has the
apeture built in).  Oh and some sort of "capture a bazillion images"
mode.