Stop attempting to reallocate buffers
You only need to allocate the buffers for the stream once and the
Raspberry Pi model I'm using has plenty of memory, so I'm not worrying
about tearing down the allocations each time.  I don't know how I've
managed to introduce a huge lag in the system on start, so that's a fun
thing I'll need to take a look at.