Theoretically implemented night mode
This should switch the display to only use the red channel for the
image.  This will naturally dim it somewhat, too.

A better approach would be to calculate the luminance of the pixel and
use that for the red channel, but meh.  I could possibly even average
the colour channels to get a reasonable approximation.  I'll see how
this goes, first.